+++ src/pipx/main.py
@@ -221,6 +221,7 @@ def run_pipx_command(args: argparse.Namespace) -> ExitCode:  # noqa: C901
             venv_args,
             verbose,
             force=args.force,
+            reinstall=False,
             include_dependencies=args.include_deps,
             suffix=args.suffix,
         )
@@ -352,7 +353,7 @@ def _add_install(subparsers: argparse._SubParsersAction) -> None:
     )
     p.add_argument(
         "--python",
-        default=DEFAULT_PYTHON,
+        # Don't pass a default Python here so we know whether --python and --force were passed at the same time
         help=(
             "The Python executable used to create the Virtual Environment and run the "
             "associated app/apps. Must be v3.6+."